Post Profile

+++ app/Http/Controllers/PostController.php
@@ -43,11 +43,8 @@ class PostController extends Controller
 
     public function destroy (Post $post)
     {
-        if(!$post->ownedBy(auth()->user()))
-        {
-            dd('no');
-        }
 
+        $this->authorize('delete' , $post);
         $post->delete();
         return back();
     }
